Fixed render issue when back to no curvature mode
Longer commit message: "Fixed rendering issue when switching back to no
curvature display mode".

When selecting 'none' in the curvature group box after having selected
one of the other curvatures, the mesh wasn't displayed correctly. The
colour buffer created in the mesh was interfering. This was the state
of the memory when displaying the curvature:

vertices	0
colours		1
normals		2

All these data were used by the 'curvature' shader, whose attributes are
on the 0,1,2 locations. When switching to the 'simple' shader, the state
of the memory didn't change (because the code for it wasn't there), and
therefore the values of the 'simple' shader, that are on the locations
0,1 took the values of the previous state of memory. This means that
when using the 'simple' shader the values in the "normals" attribute
weren't normals: they were the colours that were stored in memory in
order to display the curvature.

In order to fix this, the colour buffer is now created after all the
other buffers (see triangle_mesh.cpp). The location of the colour buffer
also changes (from 1 to 2, see src/shaders/curvature.vert).