Improved global smoothing method
Use a smaller sparse system matrix (instead of 3V x 3V, for V number of
variable vertices build one system matrix of dimensions V x V). This
means that three independent term vectors had to be used. Also, the
computation of the matrix decomposition is done on this new smaller
matrix and reused for the other vectors.