Optimised compilation of tests. Added test
Removed the header containing the prototype of all functions: whenever a
new test is added, that file has to be modified, and all the others are
recompiled.

Added new tests for inspecting a mesh. This is going to be used to test
the information contained in the edges (previous and next vertices, left
and right faces).